Extract common-build crate
