+++ common-build/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "libyal-rs-common-build"
+description = "Internal build utilities"
+license = "LGPL-3.0+"
+
+readme = "README.md"
+version = "0.1.3"
+authors = ["Omer Ben-Amram <omerbenamram@gmail.com>"]
+edition = "2018"
+
+
+[dependencies]
+failure = "0.1.5"
+bindgen = "*"
+
+[target.'cfg(target_os = "windows")'.dependencies]
+cc = "*"
+encoding_rs = "*"
+encoding_rs_io = "*"
+walkdir = "*"
